Use new Safir support for Kafka testing
Use the Kafka test container from Safir instead of including a copy
of the Safir test machinery. Move the full Kafka test into its own
directory with its own separate `conftest.py` file for fixtures, and
rename it since it no longer tests shutdown handling.
